Move ScyllaDBAPIStatus probes out of the sidecar package
